- implemented typescript - added commands: start development process and production build
