+++ week-1-ts-ecosystem/w01e02-enter-the-world-of-typescript/index.html
@@ -15,6 +15,6 @@
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="src/main.js"></script>
+    <script type="module" src="src/main.ts"></script>
   </body>
 </html>
