+++ week-1-ts-ecosystem/w01e02-enter-the-world-of-typescript/package.json
@@ -1,5 +1,12 @@
 {
   "name": "w01e02-enter-the-world-of-typescript",
   "version": "1.0.0",
-  "scripts": {}
+  "scripts": {
+    "dev": "parcel index.html",
+    "build": "parcel build index.html"
+  },
+  "devDependencies": {
+    "parcel": "^2.7.0",
+    "typescript": "^4.8.4"
+  }
 }
