--- week-1-ts-ecosystem/w01e02-enter-the-world-of-typescript/src/main.js
@@ -1,5 +1,0 @@
-import { root } from './shop/root.component.js'
-
-const $app = document.querySelector('#app')
-
-$app.append(root())
