+++ week-1-ts-ecosystem/w01e02-enter-the-world-of-typescript/src/main.ts
@@ -1,0 +1,5 @@
+import { root } from './shop/root.component'
+
+const $app = document.querySelector('#app')
+
+$app?.append(root())
