--- week-1-ts-ecosystem/w01e02-enter-the-world-of-typescript/src/shop/cart-item-sum.component.js
@@ -1,7 +1,0 @@
-import { div } from '../framework/dom-creators.js'
-
-export function cartItemSum({ value, currency = 'PLN' }) {
-  const $panelBlock = div('panel-block is-justify-content-end')
-  $panelBlock.innerHTML = `Total price: ${value} ${currency}`
-  return $panelBlock
-}
