+++ week-1-ts-ecosystem/w01e02-enter-the-world-of-typescript/src/shop/cart-item-sum.component.ts
@@ -1,0 +1,8 @@
+import {div} from '../framework/dom-creators'
+import {Price} from "../types";
+
+export function cartItemSum({value, currency = 'PLN'}: Price) {
+	const $panelBlock = div('panel-block is-justify-content-end')
+	$panelBlock.innerHTML = `Total price: ${value} ${currency}`
+	return $panelBlock
+}
