--- week-1-ts-ecosystem/w01e02-enter-the-world-of-typescript/src/shop/hero.component.js
@@ -1,13 +1,0 @@
-import { div, p, section } from '../framework/dom-creators.js'
-
-export function hero({ title, subTitle }) {
-  const $heroSection = section('hero is-link')
-  const $heroBody = div('hero-body')
-  const $title = p('title')
-  $title.innerText = title
-  const $subTitle = p('subtitle')
-  $subTitle.innerText = subTitle
-  $heroBody.append($title, $subTitle)
-  $heroSection.append($heroBody)
-  return $heroSection
-}
