+++ week-1-ts-ecosystem/w01e02-enter-the-world-of-typescript/src/shop/hero.component.ts
@@ -1,0 +1,15 @@
+import {div, p, section} from '../framework/dom-creators'
+
+type HeroProps = Record<'title' | 'subTitle', string>
+
+export function hero({title, subTitle}: HeroProps) {
+	const $heroSection = section('hero is-link')
+	const $heroBody = div('hero-body')
+	const $title = p('title')
+	$title.innerText = title
+	const $subTitle = p('subtitle')
+	$subTitle.innerText = subTitle
+	$heroBody.append($title, $subTitle)
+	$heroSection.append($heroBody)
+	return $heroSection
+}
