--- week-1-ts-ecosystem/w01e02-enter-the-world-of-typescript/src/shop/root.component.js
@@ -1,25 +1,0 @@
-import { div } from '../framework/dom-creators.js'
-import { cartPanel } from './cart-panel.component.js'
-import { hero } from './hero.component.js'
-
-const items = [
-  {
-    name: 'Tomatoes',
-    amount: 2,
-    unit: 'kg',
-    price: { value: 20, currency: 'PLN' },
-  },
-  {
-    name: 'Banana',
-    amount: 5,
-    unit: 'kg',
-    price: { value: 36, currency: 'PLN' },
-  },
-]
-
-export function root() {
-  const $hero = hero({ title: 'Shopping App', subTitle: 'buy and sell' })
-  const $container = div('container')
-  $container.append($hero, cartPanel({ items }))
-  return $container
-}
