+++ week-1-ts-ecosystem/w01e02-enter-the-world-of-typescript/src/types.d.ts
@@ -1,0 +1,2 @@
+export type Price = {value: number, currency: string}
+export type CartItem = { name: string, amount: number, unit: string, price: Price }
